Removed Windows test restrictions
