+++ board_example_test.go
@@ -1,5 +1,3 @@
-// +build !windows
-
 package glip_test
 
 import (
@@ -7,16 +5,16 @@ import (
 	"github.com/steven-xie/glip"
 )
 
-func Example() {
+func ExampleBoard() {
+	// Make a new glip.Board instance.
+	b, _ := glip.NewBoard()
+
 	// Write a string into the clipboard.
-	glip.WriteString("example string")
+	b.WriteString("example string")
 
 	// Read the string back from the clipboard. We expect it to be the same as
 	// the input string.
-	out, err := glip.ReadString()
-	if err != nil {
-		fmt.Printf("Encountered an error while reading from clipboard: %v", err)
-	}
+	out, _ := b.ReadString()
 
 	fmt.Println(out)
 	// Output: example string
